릴리즈 빌드시 네비게이션바 아이템 추가 로직 작동이 이상함. 릴리즈 빌드 후 테스트 필요.

+++ lib/component/Option_Tts.dart
@@ -205,7 +205,8 @@ class Option_Tts extends OptionsBase {
                                                 .color,
                                           ),
                                           value: (controller.config['tts']
-                                                  as Map)['groupcnt'] ??
+                                                      as Map)['groupcnt']
+                                                  .toDouble() ??
                                               1,
                                           onChanged: (value) {
                                             (controller.config['tts']
